refactor: rearranges notifier folders structure
