+++ lib/sources/presentation/notifiers/converter/connectivity_notifier.dart
