+++ lib/sources/presentation/notifiers/palettes/palette_suggestions_notifier.dart
